refactor: Extract QuestionScreen, use GameProvider
