+++ lib/presenter/question_screen.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:pub_puzzler/domain/question.dart';
+import 'package:pub_puzzler/infra/services/game_provider.dart';
+import 'package:pub_puzzler/presenter/question_card.dart';
+
+class QuestionScreen extends StatefulWidget {
+  const QuestionScreen({
+    super.key,
+    required this.category,
+    required this.difficulty,
+  });
+
+  final Category category;
+  final Difficulty difficulty;
+
+  @override
+  State<QuestionScreen> createState() => _QuestionScreenState();
+}
+
+class _QuestionScreenState extends State<QuestionScreen> {
+  void test() {}
+
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<GameProvider>(
+      builder: (context, game, child) => Scaffold(
+      //return Scaffold(
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Current Score: ${game.score}',
+              //'Current Score',
+              style: Theme.of(context).textTheme.bodyMedium!.copyWith(
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            QuestionCard(gameProvider: game, category: widget.category, difficulty: widget.difficulty),
+          ],
+        ),
+      )
+    );
+  }
+}
